Split `DirectoryInfo` tests
