+++ Tests/Testably.Abstractions.Tests/FileSystem/DirectoryInfo/CreateSubdirectory/DirectoryInfoCreateSubdirectoryTests.cs
@@ -1,8 +1,10 @@
 using Testably.Abstractions.FileSystem;
 
-namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo;
+namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo.CreateSubdirectory;
 
-public abstract partial class FileSystemDirectoryInfoTests<TFileSystem>
+// ReSharper disable once PartialTypeWithSinglePart
+public abstract partial class DirectoryInfoCreateSubdirectoryTests<TFileSystem>
+	: FileSystemTestBase<TFileSystem>
 	where TFileSystem : IFileSystem
 {
 	[SkippableTheory]
