--- Tests/Testably.Abstractions.Tests/FileSystem/DirectoryInfo/MockFileSystemTests.cs
@@ -1,33 +1,0 @@
-using Testably.Abstractions.Testing.FileSystemInitializer;
-
-namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo;
-
-// ReSharper disable once UnusedMember.Global
-public sealed class MockFileSystemTests
-	: FileSystemDirectoryInfoTests<MockFileSystem>, IDisposable
-{
-	/// <inheritdoc cref="FileSystemDirectoryInfoTests{TFileSystem}.BasePath" />
-	public override string BasePath => _directoryCleaner.BasePath;
-
-	private readonly IDirectoryCleaner _directoryCleaner;
-
-	public MockFileSystemTests() : this(new MockFileSystem())
-	{
-	}
-
-	private MockFileSystemTests(MockFileSystem mockFileSystem) : base(
-		mockFileSystem,
-		mockFileSystem.TimeSystem)
-	{
-		_directoryCleaner = FileSystem
-		   .SetCurrentDirectoryToEmptyTemporaryDirectory();
-	}
-
-	#region IDisposable Members
-
-	/// <inheritdoc cref="IDisposable.Dispose()" />
-	public void Dispose()
-		=> _directoryCleaner.Dispose();
-
-	#endregion
-}
