--- Tests/Testably.Abstractions.Tests/FileSystem/DirectoryInfo/RealFileSystemTests.cs
@@ -1,34 +1,0 @@
-#if !DEBUG || !DISABLE_TESTS_REALFILESYSTEM
-
-using Testably.Abstractions.Testing.FileSystemInitializer;
-using Xunit.Abstractions;
-
-namespace Testably.Abstractions.Tests.FileSystem.DirectoryInfo;
-
-// ReSharper disable once UnusedMember.Global
-[Collection(nameof(DriveInfoFactory.RealFileSystemTests))]
-public sealed class RealFileSystemTests :
-	FileSystemDirectoryInfoTests<RealFileSystem>,
-	IDisposable
-{
-	/// <inheritdoc cref="FileSystemDirectoryInfoTests{TFileSystem}.BasePath" />
-	public override string BasePath => _directoryCleaner.BasePath;
-
-	private readonly IDirectoryCleaner _directoryCleaner;
-
-	public RealFileSystemTests(ITestOutputHelper testOutputHelper)
-		: base(new RealFileSystem(), new RealTimeSystem())
-	{
-		_directoryCleaner = FileSystem
-		   .SetCurrentDirectoryToEmptyTemporaryDirectory(testOutputHelper.WriteLine);
-	}
-
-	#region IDisposable Members
-
-	/// <inheritdoc cref="IDisposable.Dispose()" />
-	public void Dispose()
-		=> _directoryCleaner.Dispose();
-
-	#endregion
-}
-#endif
